Implemented Wallet class for wrapping KeyringContainer
